cut copy complete

+++ src/KeyboradActions/CutAction.java
@@ -6,6 +6,7 @@ import TableCompnent.TableComponent;
 import javax.swing.*;
 import java.awt.event.ActionEvent;
 import java.io.IOException;
+import java.util.Vector;
 
 public class CutAction extends AbstractAction {
     private TableComponent table;
@@ -23,6 +24,8 @@ public class CutAction extends AbstractAction {
             // Если выделена одна ячейка
             int row = selectedRows[0];
             int col = selectedColumns[0];
+            if (col == 0)
+                return;
 
             int rowInFile = Integer.parseInt(table.tableModel.getDataVector().get(row).get(0).toString()) - 1;
             int posInRow = col - 1;
@@ -42,19 +45,19 @@ public class CutAction extends AbstractAction {
 
             System.out.println("Single cell selected at ()");
         } else if (selectedRows.length > 0 && selectedColumns.length > 0) {
-            // Если выделено несколько ячеек
             int firstRow = selectedRows[0];
             int endRow = selectedRows[selectedRows.length - 1];
             int firstCol = selectedColumns[0];
             int endCol = selectedColumns[selectedColumns.length - 1];
+            if (firstCol == 0 || endCol == 0)
+                return;
 
             int startRowInFile = Integer.parseInt(table.tableModel.getDataVector().get(firstRow).get(0).toString()) - 1;
             int startColInFile = firstCol - 1;
             int endRowInFile = Integer.parseInt(table.tableModel.getDataVector().get(endRow).get(0).toString()) - 1;
-            int endColInFile = endCol - 1;
 
             int position = table.fs.getPositionByRowCol(startRowInFile, startColInFile);
-            int cutLength = getLengthSelectBlock(startRowInFile, firstCol, endRowInFile, endCol);
+            int cutLength = table.tableModel.getLengthSelectBlock(startRowInFile, firstCol, endRowInFile, endCol);
             table.fs.removeCutBytesInFile(position, cutLength);
 
             int blockPos = TableBlock.currentBlockPos;
@@ -69,35 +72,5 @@ public class CutAction extends AbstractAction {
             // Если ячейки не выделены
             System.out.println("No cells selected");
         }
-    }
-
-    public int getLengthSelectBlock(int startRow, int startCol, int endRow, int endCol){
-        var data = table.tableModel.getDataVector();
-        int countRes = 0;
-
-        if (startRow < endRow)
-            for(int i = startRow; i <= startRow; ++i)
-                for(int j = startCol; j < data.get(i).size(); ++j){
-                    var el = data.get(i).get(j);
-                    if (el != null)
-                        countRes += 1;
-                }
-
-
-        for(int i = startRow + 1; i < endRow; ++i)
-            for(int j = 1; j < data.get(i).size(); ++j){
-                var el = data.get(i).get(j);
-                if (el != null)
-                    countRes += 1;
-            }
-
-        for(int i = endRow; i <= endRow; ++i)
-            for(int j = 1; j <= endCol; ++j){
-                var el = data.get(i).get(j);
-                if (el != null)
-                    countRes += 1;
-            }
-
-        return countRes;
     }
 }
